requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiohttp=3.8.3=py310h80987f9_0
aiosignal=1.2.0=pyhd3eb1b0_0
appnope=0.1.2=py310hca03da5_1001
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=py310hca03da5_0
attrs=22.1.0=py310hca03da5_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=openblas
bottleneck=1.3.5=py310h96f19d2_0
brotli=1.0.9=h1a28f6b_7
brotli-bin=1.0.9=h1a28f6b_7
brotlipy=0.7.0=py310h1a28f6b_1002
bzip2=1.0.8=h620ffc9_4
ca-certificates=2023.08.22=hca03da5_0
certifi=2023.7.22=py310hca03da5_0
cffi=1.15.1=py310h80987f9_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=py310hca03da5_0
comm=0.1.2=py310hca03da5_0
contourpy=1.0.5=py310h525c30c_0
cryptography=41.0.2=py310h6204c90_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.7=py310h313beb8_0
decorator=5.1.1=pyhd3eb1b0_0
et_xmlfile=1.1.0=py310hca03da5_0
executing=0.8.3=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetds=1.3.18=h02f6b3c_0
freetype=2.12.1=h1192e45_0
frozenlist=1.3.3=py310h80987f9_0
giflib=5.2.1=h80987f9_3
idna=3.4=py310hca03da5_0
ipykernel=6.25.0=py310h33ce5c2_0
ipython=8.12.2=py310hca03da5_0
jedi=0.18.1=py310hca03da5_1
joblib=1.2.0=py310hca03da5_0
jpeg=9e=h80987f9_1
jupyter_client=8.1.0=py310hca03da5_0
jupyter_core=5.3.0=py310hca03da5_0
kiwisolver=1.4.4=py310h313beb8_0
krb5=1.20.1=hf3e1bf2_1
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libbrotlicommon=1.0.9=h1a28f6b_7
libbrotlidec=1.0.9=h1a28f6b_7
libbrotlienc=1.0.9=h1a28f6b_7
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_0
libedit=3.1.20221030=h80987f9_0
libffi=3.4.4=hca03da5_0
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libiconv=1.16=h1a28f6b_2
libopenblas=0.3.21=h269037a_0
libpng=1.6.39=h80987f9_0
libsodium=1.0.18=h1a28f6b_0
libtiff=4.5.1=h313beb8_0
libwebp=1.2.4=ha3663a8_1
libwebp-base=1.2.4=h80987f9_1
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_0
matplotlib-base=3.7.2=py310h46d7db6_0
matplotlib-inline=0.1.6=py310hca03da5_0
multidict=6.0.2=py310h1a28f6b_0
munkres=1.1.4=py_0
mysql-connector-python=8.1.0=pypi_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.5.6=py310hca03da5_0
numexpr=2.8.4=py310hecc3335_1
numpy=1.25.2=py310h3b2db8e_0
numpy-base=1.25.2=py310ha9811e2_0
openai=0.27.4=py310hca03da5_0
openpyxl=3.0.10=py310h1a28f6b_0
openssl=3.0.10=h1a28f6b_2
packaging=23.1=py310hca03da5_0
pandas=2.0.3=py310h46d7db6_0
pandas-stubs=1.5.3.230203=py310hca03da5_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=py310h313beb8_0
pip=23.2.1=py310hca03da5_0
platformdirs=3.10.0=py310hca03da5_0
plotly=5.9.0=py310hca03da5_0
prompt-toolkit=3.0.36=py310hca03da5_0
protobuf=4.21.12=pypi_0
psutil=5.9.0=py310h1a28f6b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=py310hca03da5_1
pymssql=2.2.5=py310h1a28f6b_0
pyodbc=4.0.34=py310hc377ac9_0
pyopenssl=23.2.0=py310hca03da5_0
pyparsing=3.0.9=py310hca03da5_0
pysocks=1.7.1=py310hca03da5_0
python=3.10.12=hb885b13_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=1.0.0=pypi_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2022.7=py310hca03da5_0
pyzmq=25.1.0=py310h313beb8_0
readline=8.2=h1a28f6b_0
requests=2.31.0=py310hca03da5_0
scikit-learn=1.3.0=py310h46d7db6_0
scipy=1.11.1=py310h20cbe94_0
setuptools=68.0.0=py310hca03da5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h80987f9_0
stack_data=0.2.0=pyhd3eb1b0_0
tenacity=8.2.2=py310hca03da5_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=hb8d0fd4_0
tornado=6.3.2=py310h80987f9_0
tqdm=4.65.0=py310h33ce5c2_0
traitlets=5.7.1=py310hca03da5_0
types-pytz=2022.4.0.0=py310hca03da5_1
tzdata=2023c=h04d1e81_0
unixodbc=2.3.11=h1a28f6b_0
urllib3=1.26.16=py310hca03da5_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.38.4=py310hca03da5_0
xz=5.4.2=h80987f9_0
yarl=1.8.1=py310h1a28f6b_0
zeromq=4.3.4=hc377ac9_0
zlib=1.2.13=h5a0b063_0
zstd=1.5.5=hd90d995_0
